create Layout component, change navbar and footer props to static
