+++ components/navbar/Navbar.tsx
@@ -1,12 +1,17 @@
 import { useState } from "react";
 import Link from "next/link";
 
-type Props = {
-    homepageLinks: string[];
-};
-
-export const Navbar: React.FC<Props> = ({ homepageLinks }) => {
+export const Navbar: React.FC = () => {
     const [openNav, setOpenNav] = useState<boolean>(false);
+
+    // each link name keyword is tied to a specific component; constant is pre-defined for simplicity
+    // links point to sections on homepage, not separate pages
+    const homepageLinks: string[] = [
+        "About",
+        "Experience",
+        "Projects",
+        "Contact",
+    ];
 
     // render links to homepage sections; not generated pages
     const renderHomepageLinks = homepageLinks.map((e) => (
@@ -18,7 +23,7 @@ export const Navbar: React.FC<Props> = ({ homepageLinks }) => {
     ));
 
     return (
-        <nav className="flex flex-col justify-between items-start p-4 w-full sticky top-0 -mb-12 z-10 opacity-95 backdrop-blur-lg bg-white dark:bg-slate-900 border-b-2 lg:flex-row lg:items-center">
+        <nav className="flex flex-col justify-between items-start p-4 w-full opacity-95 backdrop-blur-lg bg-white dark:bg-slate-900 border-b-2 lg:flex-row lg:items-center">
             <div className="flex flex-row justify-between items-center w-full">
                 <Link href="/#">
                     <a>
