+++ pages/index.tsx
@@ -2,12 +2,12 @@ import { createClient, Entry } from "contentful";
 import type { NextPage } from "next";
 import Head from "next/head";
 
+import { Layout } from "../components/layout/Layout";
 import { Hero } from "../components/hero/Hero";
 import { About } from "../components/about/About";
 import { Experience } from "../components/experience/Experience";
 import { Projects } from "../components/projects/Projects";
 import { Contact } from "../components/contact/Contact";
-import { Footer } from "../components/footer/Footer";
 import { IHomepage, IHomepageFields } from "../@types/generated/contentful";
 
 export const getStaticProps = async () => {
@@ -51,32 +51,28 @@ const Home: NextPage<IHomepage> = ({ fields }) => {
         tools,
         projects,
         contact,
-        socialLinks,
         resume,
     } = fields;
 
     return (
-        <div className="dark:bg-slate-900">
+        <Layout>
             <Head>
                 <title>bbradforddesign</title>
                 <meta name="description" content="Blake Bradford" />
             </Head>
-            <main>
-                <Hero text={hero} />
-                <About text={about} photo={headshot} />
-                <Experience
-                    text={experience}
-                    languages={languages}
-                    frameworks={frameworks}
-                    databases={databases}
-                    tools={tools}
-                    resume={resume}
-                />
-                <Projects projects={projects} />
-                <Contact contact={contact} />
-            </main>
-            <Footer links={socialLinks} />
-        </div>
+            <Hero text={hero} />
+            <About text={about} photo={headshot} />
+            <Experience
+                text={experience}
+                languages={languages}
+                frameworks={frameworks}
+                databases={databases}
+                tools={tools}
+                resume={resume}
+            />
+            <Projects projects={projects} />
+            <Contact contact={contact} />
+        </Layout>
     );
 };
